finish moving controls over to inputsystem
