--- Assets/Scripts/ControlMappings.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 80bb53186c5f445799715c56bcd623b1
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
