--- Assets/Scripts/ControlMappings/MouseKeyboardControlsMapping.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 564f269a39b3d4e36af16a413e4b3ef6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
